+++ service/handlers_test.go
@@ -51,6 +51,14 @@ func (d mockDB) CreateProjectEntry(ctx context.Context, pe db.ProjectEntry) erro
 	return nil
 }
 
+func (d mockDB) CreateTokenEntry(ctx context.Context, project string, secretAccessor string) (db.TokenEntry, error) {
+	if project == "tokendberror" {
+		return db.TokenEntry{}, fmt.Errorf("token db error")
+	}
+
+	return db.TokenEntry{}, nil
+}
+
 func (d mockDB) ListTokenEntries(ctx context.Context, project string) ([]db.TokenEntry, error) {
 	if project == projectDoesNotExist {
 		return []db.TokenEntry{}, upper.ErrNoMoreRows
@@ -64,24 +72,7 @@ func (d mockDB) ListTokenEntries(ctx context.Context, project string) ([]db.Toke
 		return []db.TokenEntry{}, errors.New("error reading DB")
 	}
 
-	if project == "projectnotokens" {
-		return []db.TokenEntry{}, nil
-	}
-
-	return []db.TokenEntry{
-		{
-			CreatedAt: "2022-06-21T14:56:10.341066-07:00",
-			TokenID:   "ghi789",
-		},
-		{
-			CreatedAt: "2022-06-21T14:43:16.172896-07:00",
-			TokenID:   "def456",
-		},
-		{
-			CreatedAt: "2022-06-21T14:42:50.182037-07:00",
-			TokenID:   "abc123",
-		},
-	}, nil
+	return []db.TokenEntry{}, upper.ErrNoMoreRows
 }
 
 func (d mockDB) ReadProjectEntry(ctx context.Context, project string) (db.ProjectEntry, error) {
@@ -102,6 +93,14 @@ func (d mockDB) DeleteProjectEntry(ctx context.Context, project string) error {
 	}
 
 	return nil
+}
+
+func (d mockDB) DeleteTokenEntry(ctx context.Context, token string) error {
+	return nil
+}
+
+func (d mockDB) ReadTokenEntry(ctx context.Context, token string) (db.TokenEntry, error) {
+	return db.TokenEntry{}, nil
 }
 
 type mockGitClient struct{}
@@ -148,11 +147,23 @@ func newMockProvider(a credentials.Authorization, env env.Vars, h http.Header, f
 
 type mockCredentialsProvider struct{}
 
+func (m mockCredentialsProvider) DeleteProjectToken(projectName, tokenID string) error {
+	return nil
+}
+
+func (m mockCredentialsProvider) GetProjectToken(projectName, tokenID string) (types.ProjectToken, error) {
+	return types.ProjectToken{}, nil
+}
+
 func (m mockCredentialsProvider) GetToken() (string, error) {
 	return testPassword, nil
 }
 
 func (m mockCredentialsProvider) CreateProject(name string) (string, string, string, error) {
+	return "", "", "", nil
+}
+
+func (m mockCredentialsProvider) CreateToken(name string) (string, string, string, error) {
 	return "", "", "", nil
 }
 
@@ -216,6 +227,10 @@ func (m mockCredentialsProvider) ProjectExists(name string) (bool, error) {
 		"undeletableprojecttargets",
 		"undeletableproject",
 		"somedeletedberror",
+		"tokendberror",
+		"projectnotokens",
+		"projectreaderror",
+		"projectlisttokenserror",
 	}
 	for _, existingProjects := range existingProjects {
 		if name == existingProjects {
@@ -246,6 +261,7 @@ type test struct {
 	url        string
 	method     string
 	dbMock     *th.DBClientMock
+	cpMock     *th.CredsProviderMock
 }
 
 func TestCreateProject(t *testing.T) {
@@ -291,6 +307,48 @@ func TestCreateProject(t *testing.T) {
 			want:       http.StatusInternalServerError,
 			authHeader: adminAuthHeader,
 			url:        "/projects",
+			method:     "POST",
+		},
+	}
+	runTests(t, tests)
+}
+
+func TestCreateToken(t *testing.T) {
+	tests := []test{
+
+		{
+			name:       "can create token",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusOK,
+			respFile:   "TestCreateToken/can_create_token_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/undeletableprojecttargets/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "project does not exist",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusNotFound,
+			respFile:   "TestCreateToken/project_does_not_exist.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project1234/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "fails to create token when not admin",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusUnauthorized,
+			respFile:   "TestCreateToken/fails_to_create_token_when_not_admin_response.json",
+			authHeader: userAuthHeader,
+			url:        "/projects/undeletableprojecttargets/tokens",
+			method:     "POST",
+		},
+		{
+			name:       "token fails to create db entry",
+			req:        loadJSON(t, "TestCreateToken/request.json"),
+			want:       http.StatusInternalServerError,
+			authHeader: adminAuthHeader,
+			url:        "/projects/tokendberror/tokens",
 			method:     "POST",
 		},
 	}
@@ -734,6 +792,114 @@ func TestListWorkflows(t *testing.T) {
 	runTests(t, tests)
 }
 
+func TestDeleteToken(t *testing.T) {
+	tests := []test{
+		{
+			name:       "fails to delete tokens when not admin",
+			want:       http.StatusUnauthorized,
+			respFile:   "TestDeleteToken/fails_to_delete_token_when_not_admin_response.json",
+			authHeader: userAuthHeader,
+			url:        "/projects/project/tokens/existingtoken",
+			method:     "DELETE",
+		},
+		{
+			name:       "can delete token",
+			want:       http.StatusOK,
+			respFile:   "TestDeleteToken/can_delete_token_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project/tokens/existingtoken",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				GetProjectTokenFunc: func(s1 string, s2 string) (types.ProjectToken, error) {
+					return types.ProjectToken{ID: "1234"}, nil
+				},
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return true, nil
+				},
+				DeleteProjectTokenFunc: func(p, t string) error {
+					return nil
+				},
+			},
+			dbMock: &th.DBClientMock{
+				DeleteTokenEntryFunc: func(ctx context.Context, token string) error {
+					return nil
+				},
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+				ReadTokenEntryFunc: func(ctx context.Context, token string) (db.TokenEntry, error) {
+					return db.TokenEntry{ProjectID: "project1", TokenID: "1234", CreatedAt: "2022-06-21T14:42:50.182037-07:00"}, nil
+				},
+			},
+		},
+		{
+			name:       "project does not exist",
+			want:       http.StatusNotFound,
+			respFile:   "TestDeleteToken/project_does_not_exist_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/projectdoesnotexist/tokens/tokendoesnotexist",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return false, nil
+				},
+			},
+		},
+		{
+			name:       "token does not exist",
+			want:       http.StatusNotFound,
+			respFile:   "TestDeleteToken/token_does_not_exist_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project/tokens/tokendoesnotexist",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				GetProjectTokenFunc: func(s1 string, s2 string) (types.ProjectToken, error) {
+					return types.ProjectToken{}, nil
+				},
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return true, nil
+				},
+			},
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+			},
+		},
+		{
+			name:       "token delete error",
+			want:       http.StatusInternalServerError,
+			respFile:   "TestDeleteToken/token_delete_error_response.json",
+			authHeader: adminAuthHeader,
+			url:        "/projects/project/tokens/deletetokenerror",
+			method:     "DELETE",
+			cpMock: &th.CredsProviderMock{
+				DeleteProjectTokenFunc: func(s1, s2 string) error {
+					return errors.New("error deleting token from Vault")
+				},
+				GetProjectTokenFunc: func(s1 string, s2 string) (types.ProjectToken, error) {
+					return types.ProjectToken{ID: "1234"}, nil
+				},
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return true, nil
+				},
+			},
+			dbMock: &th.DBClientMock{
+				DeleteTokenEntryFunc: func(ctx context.Context, token string) error {
+					return errors.New("error deleting entry from DB")
+				},
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+				ReadTokenEntryFunc: func(ctx context.Context, token string) (db.TokenEntry, error) {
+					return db.TokenEntry{ProjectID: "project1", TokenID: "1234", CreatedAt: "2022-06-21T14:42:50.182037-07:00"}, nil
+				},
+			},
+		},
+	}
+	runTestsV2(t, tests)
+}
+
 func TestListTokens(t *testing.T) {
 	tests := []test{
 		{
@@ -751,6 +917,27 @@ func TestListTokens(t *testing.T) {
 			authHeader: adminAuthHeader,
 			url:        "/projects/undeletableprojecttargets/tokens",
 			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, p string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1", Repository: "repo"}, nil
+				},
+				ListTokenEntriesFunc: func(ctx context.Context, project string) ([]db.TokenEntry, error) {
+					return []db.TokenEntry{
+						{
+							CreatedAt: "2022-06-21T14:56:10.341066-07:00",
+							TokenID:   "ghi789",
+						},
+						{
+							CreatedAt: "2022-06-21T14:43:16.172896-07:00",
+							TokenID:   "def456",
+						},
+						{
+							CreatedAt: "2022-06-21T14:42:50.182037-07:00",
+							TokenID:   "abc123",
+						},
+					}, nil
+				},
+			},
 		},
 		{
 			name:       "project not found",
@@ -766,20 +953,20 @@ func TestListTokens(t *testing.T) {
 			},
 		},
 		{
-			name:       "project not found",
-			want:       http.StatusNotFound,
-			respFile:   "TestListTokens/project_not_found_response.json",
-			authHeader: adminAuthHeader,
-			url:        "/projects/projectdoesnotexist/tokens",
-			method:     "GET",
-		},
-		{
 			name:       "no tokens",
 			want:       http.StatusOK,
 			respFile:   "TestListTokens/no_tokens_response.json",
 			authHeader: adminAuthHeader,
 			url:        "/projects/projectnotokens/tokens",
 			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, p string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "abc123", Repository: "repo"}, nil
+				},
+				ListTokenEntriesFunc: func(ctx context.Context, project string) ([]db.TokenEntry, error) {
+					return []db.TokenEntry{}, nil
+				},
+			},
 		},
 		{
 			name:       "project read error",
@@ -788,6 +975,11 @@ func TestListTokens(t *testing.T) {
 			authHeader: adminAuthHeader,
 			url:        "/projects/projectreaderror/tokens",
 			method:     "GET",
+			cpMock: &th.CredsProviderMock{
+				ProjectExistsFunc: func(s string) (bool, error) {
+					return false, errors.New("error retrieving project")
+				},
+			},
 		},
 		{
 			name:       "list tokens read error",
@@ -796,9 +988,16 @@ func TestListTokens(t *testing.T) {
 			authHeader: adminAuthHeader,
 			url:        "/projects/projectlisttokenserror/tokens",
 			method:     "GET",
+			dbMock: &th.DBClientMock{
+				ReadProjectEntryFunc: func(ctx context.Context, project string) (db.ProjectEntry, error) {
+					return db.ProjectEntry{ProjectID: "project1"}, nil
+				},
+				ListTokenEntriesFunc: func(ctx context.Context, project string) ([]db.TokenEntry, error) {
+					return []db.TokenEntry{}, errors.New("error from DB")
+				},
+			},
 		},
 	}
-
 	runTestsV2(t, tests)
 }
 
@@ -938,13 +1137,19 @@ func runTests(t *testing.T, tests []test) {
 
 				defer resp.Body.Close()
 
-				assert.JSONEq(t, string(wantBody), string(body))
+				bodyStr := string(body)
+				wantBodyStr := string(wantBody)
+
+				if bodyStr == "" && wantBodyStr == "" {
+					assert.Equal(t, bodyStr, wantBodyStr)
+				} else {
+					assert.JSONEq(t, bodyStr, wantBodyStr)
+				}
 			}
 		})
 	}
 }
 
-// Run tests, checking the response status codes.
 func runTestsV2(t *testing.T, tests []test) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -968,6 +1173,15 @@ func runTestsV2(t *testing.T, tests []test) {
 
 			if tt.dbMock != nil {
 				h.dbClient = tt.dbMock
+			}
+
+			if tt.cpMock != nil {
+				defaultCPFunc := func(a credentials.Authorization, env env.Vars, h http.Header, f credentials.VaultConfigFn, fn credentials.VaultSvcFn) (credentials.Provider, error) {
+					// TODO: probably best to create a base/default CP mock struct here
+					return tt.cpMock, nil
+				}
+
+				h.newCredentialsProvider = defaultCPFunc
 			}
 
 			resp := executeRequestWithHandler(h, tt.method, tt.url, serialize(tt.req), tt.authHeader)
@@ -997,7 +1211,15 @@ func runTestsV2(t *testing.T, tests []test) {
 
 				defer resp.Body.Close()
 
-				assert.JSONEq(t, string(wantBody), string(body))
+				bodyStr := string(body)
+				wantBodyStr := string(wantBody)
+
+				// don't use assert.JSONEq for empty strings
+				if bodyStr == "" && wantBodyStr == "" {
+					assert.Equal(t, bodyStr, wantBodyStr)
+				} else {
+					assert.JSONEq(t, wantBodyStr, bodyStr)
+				}
 			}
 		})
 	}